adding an allert before deleting an item

+++ src/components/ProfileContentCard/ProfileContentCard.jsx
@@ -2,14 +2,14 @@ import React from 'react'
 
 import moment from 'moment'
 import { useHistory } from 'react-router-dom'
-
+import { useTranslation } from 'react-i18next'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import PropTypes from 'prop-types'
 import { db } from '../../firebase'
 
 export default function ProfileContentCard({ title, content, seen, time, id }) {
   const history = useHistory()
-
+  const { t } = useTranslation()
   function timing(date) {
     const now = moment()
     const then = moment(date)
@@ -22,10 +22,12 @@ export default function ProfileContentCard({ title, content, seen, time, id }) {
   }
 
   const deleteItem = productId => {
-    db.collection('products')
-      .doc(productId)
-      .delete()
-      .then(() => window.location.reload(false))
+    if (window.confirm(t('profile.deleting'))) {
+      db.collection('products')
+        .doc(productId)
+        .delete()
+        .then(() => window.location.reload(false))
+    }
   }
   const handleClick = () => {
     history.push(`/productdetails/${id}`)
@@ -60,7 +62,9 @@ export default function ProfileContentCard({ title, content, seen, time, id }) {
           <button
             type="button"
             className="bg-red mx-2 px-4 py-1 rounded-full text-white hover:shadow-none focus:outline-none shadow-md  transition duration-300 ease-in-out "
-            onClick={() => deleteItem(id)}
+            onClick={() => {
+              deleteItem(id)
+            }}
           >
             <FontAwesomeIcon icon="trash-alt" />
           </button>
